+++ fj-doc-maven-plugin/src/main/java/org/fugerit/java/doc/project/facade/FlavourContext.java
@@ -38,6 +38,12 @@ public class FlavourContext {
     @Getter @Setter
     private String flavourVersion;
 
+    @Getter @Setter
+    private String version;
+
+    @Getter @Setter
+    private String extensions;
+
     private String toClassName( String base, String splitString ) {
         StringBuilder buf = new StringBuilder();
         String[] split = base.split( splitString );
@@ -50,6 +56,10 @@ public class FlavourContext {
     public String getArtifactIdAsClassName() {
         String res = toClassName( this.artifactId, "-" );
         return toClassName( res, "\\." );
+    }
+
+    public boolean isAsciidocFreemarkerHandlerAvailable() {
+        return VersionCheck.isMajorThan( VenusContext.VERSION_ASCIIDOC_FREEMARKER_HANDLER, this.getVersion() );
     }
 
 }